[CatalogPromotion] Move scope validator and constraint to PromotionBundle
