+++ src/Sylius/Bundle/CoreBundle/CatalogPromotion/Validator/CatalogPromotionAction/FixedDiscountActionValidator.php
@@ -13,8 +13,6 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\CatalogPromotion\Validator\CatalogPromotionAction;
 
-use Sylius\Bundle\ApiBundle\SectionResolver\AdminApiSection;
-use Sylius\Bundle\CoreBundle\SectionResolver\SectionProviderInterface;
 use Sylius\Bundle\PromotionBundle\Validator\CatalogPromotionAction\ActionValidatorInterface;
 use Sylius\Bundle\PromotionBundle\Validator\Constraints\CatalogPromotionAction;
 use Sylius\Component\Channel\Repository\ChannelRepositoryInterface;
