+++ src/Sylius/Bundle/CoreBundle/CatalogPromotion/Validator/CatalogPromotionScope/ForTaxonsScopeValidator.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\CatalogPromotion\Validator\CatalogPromotionScope;
 
-use Sylius\Bundle\CoreBundle\CatalogPromotion\Validator\Constraints\CatalogPromotionScope;
 use Sylius\Bundle\PromotionBundle\Validator\CatalogPromotionScope\ScopeValidatorInterface;
+use Sylius\Bundle\PromotionBundle\Validator\Constraints\CatalogPromotionScope;
 use Sylius\Component\Taxonomy\Repository\TaxonRepositoryInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
@@ -33,7 +33,11 @@ final class ForTaxonsScopeValidator implements ScopeValidatorInterface
 
         foreach ($configuration['taxons'] as $taxonCode) {
             if (null === $this->taxonRepository->findOneBy(['code' => $taxonCode])) {
-                $context->buildViolation($constraint->invalidTaxons)->atPath('configuration.taxons')->addViolation();
+                $context
+                    ->buildViolation('sylius.catalog_promotion_scope.for_taxons.invalid_taxons')
+                    ->atPath('configuration.taxons')
+                    ->addViolation()
+                ;
 
                 return;
             }
