+++ src/Sylius/Bundle/CoreBundle/Resources/config/validation/CatalogPromotionScope.xml
@@ -13,7 +13,7 @@
 
 <constraint-mapping xmlns="http://symfony.com/schema/dic/constraint-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping http://symfony.com/schema/dic/services/constraint-mapping-1.0.xsd">
     <class name="Sylius\Component\Core\Model\CatalogPromotionScope">
-        <constraint name="Sylius\Bundle\CoreBundle\CatalogPromotion\Validator\Constraints\CatalogPromotionScope">
+        <constraint name="Sylius\Bundle\PromotionBundle\Validator\Constraints\CatalogPromotionScope">
             <option name="groups">sylius</option>
         </constraint>
     </class>
